@@ -9,7 +9,10 @@
     "build-prod": "ng build --configuration production",
     "watch": "ng build --watch --configuration development",
     "build:ssr": "ng run ScriptBeeWebApp:server:dev",
-    "test": "ng test",
+    "test": "jest",
+    "test:watch": "jest --watch",
+    "test:pact": "jest --testTimeout 30000 --testPathPattern .*.pact.spec.ts",
+    "clean:pact": "rimraf ./pacts",
     "lint": "ng lint",
     "e2e": "ng e2e",
     "run:all": "node node_modules/@angular-architects/module-federation/src/server/mf-dev-server.js"
@@ -54,22 +57,20 @@
     "@angular/cli": "14.2.2",
     "@angular/compiler-cli": "14.2.1",
     "@ngneat/spectator": "^6.0.0",
-    "@types/jasmine": "^4.3.0",
+    "@pact-foundation/pact": "^10.4.1",
+    "@types/jest": "^29.4.0",
     "@types/node": "18.7.16",
     "@typescript-eslint/eslint-plugin": "5.36.2",
     "@typescript-eslint/parser": "5.36.2",
-    "eslint": "^8.23.0",
-    "eslint-config-prettier": "^8.5.0",
+    "eslint": "^8.34.0",
+    "eslint-config-prettier": "^8.6.0",
     "eslint-plugin-prettier": "^4.2.1",
-    "jasmine-core": "^4.4.0",
-    "karma": "~6.4.0",
-    "karma-chrome-launcher": "~3.1.1",
-    "karma-coverage": "^2.2.0",
-    "karma-jasmine": "^5.1.0",
-    "karma-jasmine-html-reporter": "2.0.0",
+    "jest": "^28.1.3",
+    "jest-preset-angular": "^12.2.6",
     "ng-mocks": "^14.5.0",
     "ngx-build-plus": "^14.0.0",
     "prettier": "^2.8.1",
+    "rimraf": "^4.1.2",
     "typescript": "~4.7.0"
   }
 }
