@@ -1,15 +1,14 @@
-import { Injectable } from '@angular/core';
-import { HttpClient } from "@angular/common/http";
-import { SignalrClient, SignalrConnection } from "ngx-signalr-websocket";
-import { ReplaySubject } from "rxjs";
-import { WatchedFile } from "./watchedFile";
-import { map } from "rxjs/operators";
+import {Injectable} from '@angular/core';
+import {HttpClient} from '@angular/common/http';
+import {SignalrClient, SignalrConnection} from 'ngx-signalr-websocket';
+import {ReplaySubject} from 'rxjs';
+import {WatchedFile} from './watchedFile';
+import {map} from 'rxjs/operators';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class NotificationsService {
-
   private connection?: SignalrConnection;
 
   watchedFiles = new ReplaySubject<WatchedFile>();
@@ -17,33 +16,28 @@ export class NotificationsService {
   constructor(httpClient: HttpClient) {
     const client = SignalrClient.create(httpClient);
 
-    client.connect('/api/fileWatcherHub')
-      .subscribe({
-        next: connection => {
-          console.log(connection)
-          this.connection = connection;
-          this.setupWatchedFiles(connection);
-        },
-        error: err => console.error('SignalR Connection Error: ', err)
-      });
+    client.connect('/api/fileWatcherHub').subscribe({
+      next: (connection) => {
+        this.connection = connection;
+        this.setupWatchedFiles(connection);
+      },
+      error: (err) => console.error('SignalR Connection Error: ', err),
+    });
   }
 
   send() {
-    console.log(this.connection)
     this.connection?.invoke('SendFileWatch', {
       path: 'p',
-      content: 'c'
-    })
+      content: 'c',
+    });
   }
 
   private setupWatchedFiles(connection: SignalrConnection) {
-    connection.on<[WatchedFile]>('ReceiveFileWatch')
-      .pipe(
-        map(value => value[0]),
-      )
-      .subscribe(value => {
+    connection
+      .on<[WatchedFile]>('ReceiveFileWatch')
+      .pipe(map((value) => value[0]))
+      .subscribe((value) => {
         this.watchedFiles.next(value);
       });
   }
 }
-
