@@ -1,7 +1,4 @@
-﻿using System.Threading;
-using System.Threading.Tasks;
-
-namespace DxWorks.ScriptBee.Plugin.Api;
+﻿namespace DxWorks.ScriptBee.Plugin.Api;
 
 public interface IHelperFunctions : IPlugin
 {
