@@ -6,6 +6,7 @@ namespace ScriptBee.Models;
 
 public class ProjectModel : IDocument
 {
+    [BsonId] public string Id { get; set; } = null!;
     public string Name { get; set; } = "";
     public DateTime CreationDate { get; set; }
     public Dictionary<string, List<FileData>> SavedFiles { get; set; } = new();
@@ -13,5 +14,4 @@ public class ProjectModel : IDocument
     public Dictionary<string, List<FileData>> LoadedFiles { get; set; } = new();
     public string Linker { get; set; } = "";
     public Run? LastRun { get; set; }
-    [BsonId] public string Id { get; set; }
 }
