@@ -2,4 +2,4 @@
 
 namespace ScriptBee.ProjectContext;
 
-public record FileTreeNode(string name, List<FileTreeNode>? children, string filePath, string srcPath);
+public record FileTreeNode(string Name, string FilePath, string SrcPath, List<FileTreeNode>? Children);
