@@ -1,11 +1,11 @@
 import { Routes } from '@angular/router';
 import { ProjectsComponent } from './projects/projects.component';
-import { ProjectDetailsComponent } from './project-details/project-details.component';
-import { DetailsContentComponent } from './project-details/details-content/details-content.component';
-import { NoScriptsComponent } from './project-details/scripts-content/no-scripts/no-scripts.component';
-import { SelectedScriptComponent } from './project-details/scripts-content/selected-script/selected-script.component';
-import { ScriptsContentComponent } from './project-details/scripts-content/scripts-content.component';
-import { PluginsMarketplaceDashboardComponent } from './plugins/plugins-marketplace-dashboard/plugins-marketplace-dashboard.component';
+import { ProjectDetailsComponent } from './project-details/components/project-details.component';
+import { DetailsContentComponent } from './project-details/components/details-content/details-content.component';
+import { NoScriptsComponent } from './project-details/components/run-script/scripts-content/no-scripts/no-scripts.component';
+import { SelectedScriptComponent } from './project-details/components/run-script/selected-script/selected-script.component';
+import { ScriptsContentComponent } from './project-details/components/run-script/scripts-content.component';
+import { PluginsMarketplaceDashboardComponent } from './plugins/components/plugins-marketplace-dashboard/plugins-marketplace-dashboard.component';
 
 export const ROUTES: Routes = [
   // {path: '', component: HomeComponent, pathMatch: 'full'},
@@ -22,10 +22,10 @@ export const ROUTES: Routes = [
         component: ScriptsContentComponent,
         children: [
           { path: '', component: NoScriptsComponent },
-          { path: ':scriptPath', component: SelectedScriptComponent }
-        ]
+          { path: ':scriptPath', component: SelectedScriptComponent },
+        ],
       },
-      { path: '**', redirectTo: 'details' }
-    ]
-  }
+      { path: '**', redirectTo: 'details' },
+    ],
+  },
 ];
