@@ -1,16 +1,14 @@
-import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs';
-import { FileTreeNode } from '../../project-details/scripts-content/fileTreeNode';
-import { HttpClient, HttpParams } from '@angular/common/http';
-import { contentHeaders } from '../../shared/headers';
+import {Injectable} from '@angular/core';
+import {Observable} from 'rxjs';
+import {FileTreeNode} from '../../project-details/components/run-script/scripts-content/fileTreeNode';
+import {HttpClient, HttpParams} from '@angular/common/http';
+import {contentHeaders} from '../../shared/headers';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class FileSystemService {
   private projectStructureAPIUrl = '/api/projectstructure';
-  private projectStructureScriptAPIUrl = '/api/projectstructure/script';
-  private projectStructureScriptGetAbsolutePathAPIUrl = '/api/projectstructure/scriptabsolutepath';
   private projectStructureProjectAbsolutePath = '/api/projectstructure/projectabsolutepath';
   private projectStructureFileWatcherUrl = '/api/projectstructure/filewatcher';
 
@@ -18,35 +16,7 @@ export class FileSystemService {
 
   getFileSystem(projectId: string): Observable<FileTreeNode> {
     return this.http.get<FileTreeNode>(`${this.projectStructureAPIUrl}/${projectId}`, {
-      headers: contentHeaders
-    });
-  }
-
-  createScript(projectId: string, filePath: string, scriptType: string) {
-    return this.http.post<FileTreeNode>(
-      this.projectStructureScriptAPIUrl,
-      {
-        projectId,
-        filePath,
-        scriptType
-      },
-      { headers: contentHeaders }
-    );
-  }
-
-  getFileContent(projectId: string, filePath: string): Observable<string> {
-    const params = new HttpParams().set('projectId', projectId).set('filePath', filePath);
-    return this.http.get<string>(this.projectStructureScriptAPIUrl, {
       headers: contentHeaders,
-      params
-    });
-  }
-
-  getScriptAbsolutePath(projectId: string, filePath: string): Observable<string> {
-    const params = new HttpParams().set('projectId', projectId).set('filePath', filePath);
-    return this.http.get<string>(this.projectStructureScriptGetAbsolutePathAPIUrl, {
-      headers: contentHeaders,
-      params
     });
   }
 
@@ -54,7 +24,7 @@ export class FileSystemService {
     const params = new HttpParams().set('projectId', projectId);
     return this.http.get<string>(this.projectStructureProjectAbsolutePath, {
       headers: contentHeaders,
-      params
+      params,
     });
   }
 
@@ -64,7 +34,7 @@ export class FileSystemService {
 
   removeFileWatcher(projectId: string) {
     return this.http.delete(`${this.projectStructureFileWatcherUrl}/${projectId}`, {
-      headers: contentHeaders
+      headers: contentHeaders,
     });
   }
 }
