@@ -1,22 +1,18 @@
 import { Injectable } from '@angular/core';
-import {
-  ErrorDialogComponent
-} from "../../project-details/details-content/error-dialog/error-dialog/error-dialog.component";
-import { MatDialog } from "@angular/material/dialog";
+import { ErrorDialogComponent } from './error-dialog.component';
+import { MatDialog } from '@angular/material/dialog';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class ErrorDialogService {
-
-  constructor(private dialog: MatDialog) {
-  }
+  constructor(private dialog: MatDialog) {}
 
   displayDialogErrorMessage(mainProblem: string, errorMessage: string) {
     if (errorMessage == null || errorMessage === '') {
-      this.dialog.open(ErrorDialogComponent, {data: {mainProblem: mainProblem, errorMessage: ''}});
+      this.dialog.open(ErrorDialogComponent, { data: { mainProblem: mainProblem, errorMessage: '' } });
     } else {
-      this.dialog.open(ErrorDialogComponent, {data: {mainProblem: mainProblem, errorMessage: errorMessage}});
+      this.dialog.open(ErrorDialogComponent, { data: { mainProblem: mainProblem, errorMessage: errorMessage } });
     }
   }
 }
