@@ -1,4 +1,5 @@
-﻿using System.Threading.Tasks;
+﻿using System.Collections.Generic;
+using System.Threading.Tasks;
 using FluentValidation.TestHelper;
 using ScriptBeeWebApp.EndpointDefinitions.Arguments;
 using ScriptBeeWebApp.EndpointDefinitions.Arguments.Validation;
@@ -18,7 +19,7 @@ public class CreateScriptValidatorTests
     [Fact]
     public async Task GivenValidCreateScript_WhenValidate_ThenResultHasNoErrors()
     {
-        var createScript = new CreateScript("id1", "path", "type");
+        var createScript = new CreateScript("id1", "path", "type", new List<ScriptParameter>());
 
         var result = await _validator.TestValidateAsync(createScript);
 
@@ -28,7 +29,7 @@ public class CreateScriptValidatorTests
     [Fact]
     public async Task GivenEmptyProjectId_WhenValidate_ThenResultHasErrors()
     {
-        var createScript = new CreateScript("", "path", "type");
+        var createScript = new CreateScript("", "path", "type", new List<ScriptParameter>());
 
         var result = await _validator.TestValidateAsync(createScript);
 
@@ -38,7 +39,7 @@ public class CreateScriptValidatorTests
     [Fact]
     public async Task GivenEmptyFilePath_WhenValidate_ThenResultHasErrors()
     {
-        var createScript = new CreateScript("projectId", "", "type");
+        var createScript = new CreateScript("projectId", "", "type", new List<ScriptParameter>());
 
         var result = await _validator.TestValidateAsync(createScript);
 
@@ -48,10 +49,33 @@ public class CreateScriptValidatorTests
     [Fact]
     public async Task GivenEmptyScriptType_WhenValidate_ThenResultHasErrors()
     {
-        var createScript = new CreateScript("projectId", "path", "");
+        var createScript = new CreateScript("projectId", "path", "", new List<ScriptParameter>());
 
         var result = await _validator.TestValidateAsync(createScript);
 
-        result.ShouldHaveValidationErrorFor(r => r.ScriptType);
+        result.ShouldHaveValidationErrorFor(r => r.ScriptLanguage);
+    }
+
+    [Fact]
+    public async Task GivenEmptyParameters_WhenValidate_ThenResultHasNoErrors()
+    {
+        var createScript = new CreateScript("projectId", "path", "type", new List<ScriptParameter>());
+
+        var result = await _validator.TestValidateAsync(createScript);
+
+        result.ShouldNotHaveAnyValidationErrors();
+    }
+
+    [Fact]
+    public async Task GivenParametersWithEmptyName_WhenValidate_ThenResultHasErrors()
+    {
+        var createScript = new CreateScript("projectId", "path", "type", new List<ScriptParameter>
+        {
+            new("", "type", "value")
+        });
+
+        var result = await _validator.TestValidateAsync(createScript);
+
+        result.ShouldHaveAnyValidationError();
     }
 }
