@@ -6,12 +6,11 @@ using AutoFixture;
 using DxWorks.ScriptBee.Plugin.Api;
 using FluentValidation;
 using FluentValidation.Results;
-using Microsoft.AspNetCore.Mvc;
 using Moq;
 using ScriptBee.ProjectContext;
 using ScriptBeeWebApp.EndpointDefinitions;
 using ScriptBeeWebApp.EndpointDefinitions.Arguments;
-using ScriptBeeWebApp.EndpointDefinitions.Arguments.Validation;
+using ScriptBeeWebApp.EndpointDefinitions.DTO;
 using ScriptBeeWebApp.Services;
 using Xunit;
 
@@ -24,6 +23,7 @@ public class GenerateScriptEndpointDefinitionTests
 
     private readonly Mock<IValidator<GenerateScriptRequest>> _generateScriptRequestValidatorMock;
     private readonly Mock<IGenerateScriptService> _generateScriptServiceMock;
+
     private readonly Mock<IProjectManager> _projectManagerMock;
 
     public GenerateScriptEndpointDefinitionTests()
@@ -35,16 +35,6 @@ public class GenerateScriptEndpointDefinitionTests
         _fixture = new Fixture();
     }
 
-    [Fact]
-    public void GivenSupportedLanguages_WhenGetLanguages_ThenSupportedLanguagesAreReturned()
-    {
-        _generateScriptServiceMock.Setup(s => s.GetSupportedLanguages())
-            .Returns(new List<string> { "C#", "JavaScript" });
-
-        var languages = GenerateScriptEndpointDefinition.GetLanguages(_generateScriptServiceMock.Object);
-
-        Assert.Equal(new List<string> { "C#", "JavaScript" }, languages);
-    }
 
     [Fact]
     public async Task GivenInvalidGenerateScriptRequest_WhenPostGenerateScript_ThenBadRequestIsReturned()
@@ -59,7 +49,7 @@ public class GenerateScriptEndpointDefinitionTests
             .Setup(v => v.ValidateAsync(generateScriptRequest, It.IsAny<CancellationToken>()))
             .ReturnsAsync(new ValidationResult(new List<ValidationFailure> { new("property", "error") }));
 
-        var result = await GenerateScriptEndpointDefinition.PostGenerateScript(generateScriptRequest,
+        var result = await ScriptsEndpointDefinition.PostGenerateScript(generateScriptRequest,
             _generateScriptRequestValidatorMock.Object, _generateScriptServiceMock.Object, _projectManagerMock.Object);
 
         Assert.Equal(400, result.GetStatusCode());
@@ -77,7 +67,7 @@ public class GenerateScriptEndpointDefinitionTests
         _generateScriptServiceMock.Setup(s => s.GetGenerationStrategy("invalid"))
             .Returns((IScriptGeneratorStrategy?)null);
 
-        var result = await GenerateScriptEndpointDefinition.PostGenerateScript(generateScriptRequest,
+        var result = await ScriptsEndpointDefinition.PostGenerateScript(generateScriptRequest,
             _generateScriptRequestValidatorMock.Object, _generateScriptServiceMock.Object, _projectManagerMock.Object);
 
         Assert.Equal(400, result.GetStatusCode());
@@ -96,7 +86,7 @@ public class GenerateScriptEndpointDefinitionTests
             .Returns(new Mock<IScriptGeneratorStrategy>().Object);
         _projectManagerMock.Setup(p => p.GetProject("id1")).Returns((Project?)null);
 
-        var result = await GenerateScriptEndpointDefinition.PostGenerateScript(generateScriptRequest,
+        var result = await ScriptsEndpointDefinition.PostGenerateScript(generateScriptRequest,
             _generateScriptRequestValidatorMock.Object, _generateScriptServiceMock.Object, _projectManagerMock.Object);
 
         Assert.Equal(404, result.GetStatusCode());
@@ -122,7 +112,7 @@ public class GenerateScriptEndpointDefinitionTests
                 It.IsAny<IScriptGeneratorStrategy>(), It.IsAny<CancellationToken>()))
             .ReturnsAsync(new Mock<Stream>().Object);
 
-        var result = await GenerateScriptEndpointDefinition.PostGenerateScript(generateScriptRequest,
+        var result = await ScriptsEndpointDefinition.PostGenerateScript(generateScriptRequest,
             _generateScriptRequestValidatorMock.Object, _generateScriptServiceMock.Object, _projectManagerMock.Object);
 
         Assert.Equal("validSampleCode.zip", result.GetProperty<string>("FileDownloadName"));
