@@ -10,7 +10,8 @@ using ScriptBee.Models;
 using ScriptBee.ProjectContext;
 using ScriptBeeWebApp.EndpointDefinitions;
 using ScriptBeeWebApp.EndpointDefinitions.Arguments;
-using ScriptBeeWebApp.EndpointDefinitions.Arguments.Validation;
+using ScriptBeeWebApp.EndpointDefinitions.DTO;
+using ScriptBeeWebApp.Repository;
 using ScriptBeeWebApp.Services;
 using Xunit;
 
