@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using AutoFixture;
@@ -9,10 +8,10 @@ using DxWorks.ScriptBee.Plugin.Api;
 using Moq;
 using ScriptBee.Models;
 using ScriptBee.Plugin;
-using ScriptBee.Plugin.Manifest;
 using ScriptBee.ProjectContext;
 using ScriptBee.Services;
 using ScriptBeeWebApp.Data.Exceptions;
+using ScriptBeeWebApp.Repository;
 using ScriptBeeWebApp.Services;
 using Xunit;
 
@@ -26,6 +25,7 @@ public class RunScriptServiceTests
     private readonly Mock<IProjectFileStructureManager> _projectFileStructureManagerMock;
     private readonly Mock<IProjectModelService> _projectModelServiceMock;
     private readonly Mock<IRunModelService> _runModelServiceMock;
+    private readonly Mock<IScriptsService> _scriptsServiceMock;
     private readonly Fixture _fixture;
 
     private readonly RunScriptService _runScriptService;
@@ -38,36 +38,12 @@ public class RunScriptServiceTests
         _projectFileStructureManagerMock = new Mock<IProjectFileStructureManager>();
         _projectModelServiceMock = new Mock<IProjectModelService>();
         _runModelServiceMock = new Mock<IRunModelService>();
+        _scriptsServiceMock = new Mock<IScriptsService>();
         _fixture = new Fixture();
 
         _runScriptService = new RunScriptService(_fileModelServiceMock.Object, _guidGeneratorMock.Object,
             _pluginRepositoryMock.Object, _projectFileStructureManagerMock.Object, _projectModelServiceMock.Object,
-            _runModelServiceMock.Object);
-    }
-
-
-    [Fact]
-    public void GivenLoadedPluginManifests_WhenGetSupportedLanguages_ThenReturnSupportedLanguages()
-    {
-        var pluginExtensionPoints = new List<ScriptRunnerPluginExtensionPoint>
-        {
-            new()
-            {
-                Language = "C#"
-            },
-            new()
-            {
-                Language = "Python"
-            }
-        };
-        _pluginRepositoryMock.Setup(x => x.GetLoadedPluginExtensionPoints<ScriptRunnerPluginExtensionPoint>())
-            .Returns(pluginExtensionPoints);
-
-        var supportedLanguages = _runScriptService.GetSupportedLanguages().ToList();
-
-        Assert.Equal(2, supportedLanguages.Count);
-        Assert.Contains("C#", supportedLanguages);
-        Assert.Contains("Python", supportedLanguages);
+            _runModelServiceMock.Object, _scriptsServiceMock.Object);
     }
 
     [Fact]
