@@ -1,6 +1,6 @@
 ﻿namespace ScriptBeeWebApp.EndpointDefinitions.Arguments;
 
-public record CreateScript(string ProjectId, string FilePath, string ScriptType)
+public record CreateScript(string ProjectId, string FilePath, string ScriptLanguage, List<ScriptParameter> Parameters)
 {
     public string FilePath { get; set; } = FilePath;
 }
