@@ -12,15 +12,12 @@ public class ProjectStructureEndpointDefinition : IEndpointDefinition
 {
     public void DefineServices(IServiceCollection services)
     {
-        //
+        services.AddSingleton<IConfigFoldersService, ConfigFoldersService>();
     }
 
     public void DefineEndpoints(WebApplication app)
     {
         app.MapGet("/api/projectstructure/{projectId}", GetProjectStructure);
-        app.MapPost("/api/projectstructure/script", CreateScript);
-        app.MapGet("/api/projectstructure/script", GetScriptContent);
-        app.MapGet("/api/projectstructure/scriptabsolutepath", GetScriptAbsolutePath);
         app.MapGet("/api/projectstructure/projectabsolutepath", GetProjectAbsolutePath);
         app.MapPost("/api/projectstructure/filewatcher", SetupFileWatcher);
         app.MapDelete("/api/projectstructure/filewatcher/{projectId}", RemoveFileWatcher);
@@ -39,66 +36,7 @@ public class ProjectStructureEndpointDefinition : IEndpointDefinition
         return Results.Ok(fileTreeNode);
     }
 
-    public static async Task<IResult> CreateScript([FromBody] CreateScript createScript,
-        IValidator<CreateScript> validator, IProjectFileStructureManager projectFileStructureManager,
-        IProjectStructureService projectStructureService,
-        CancellationToken cancellationToken = default)
-    {
-        var validationResult = await validator.ValidateAsync(createScript, cancellationToken);
-        if (!validationResult.IsValid)
-        {
-            return Results.BadRequest(validationResult.GetValidationErrorsResponse());
-        }
-
-        var (extension, content) =
-            await projectStructureService.GetSampleCodeAsync(createScript.ScriptType, cancellationToken);
-
-        if (!createScript.FilePath.EndsWith(extension))
-        {
-            createScript.FilePath += extension;
-        }
-
-        if (projectFileStructureManager.FileExists(createScript.ProjectId, createScript.FilePath))
-        {
-            return Results.Conflict();
-        }
-
-        var node = projectFileStructureManager.CreateSrcFile(createScript.ProjectId, createScript.FilePath,
-            content);
-
-        return Results.Ok(node);
-    }
-
-    public static async Task<IResult> GetScriptContent([FromQuery] string projectId, [FromQuery] string filePath,
-        IProjectFileStructureManager projectFileStructureManager, CancellationToken cancellationToken)
-    {
-        if (string.IsNullOrEmpty(projectId) || string.IsNullOrEmpty(filePath))
-        {
-            return Results.BadRequest("Invalid arguments!");
-        }
-
-        var content = await projectFileStructureManager.GetFileContentAsync(projectId, filePath);
-
-        if (content == null)
-        {
-            return Results.NotFound("Script not found");
-        }
-
-        return Results.Ok(content);
-    }
-
-    public static IResult GetScriptAbsolutePath([FromQuery] string projectId, [FromQuery] string filePath,
-        IProjectFileStructureManager projectFileStructureManager, CancellationToken cancellationToken)
-    {
-        if (string.IsNullOrEmpty(projectId) || string.IsNullOrEmpty(filePath))
-        {
-            return Results.BadRequest("Invalid arguments!");
-        }
-
-        return Results.Ok(projectFileStructureManager.GetAbsoluteFilePath(projectId, filePath));
-    }
-
-    public static IResult GetProjectAbsolutePath([FromQuery] string projectId,
+    private static IResult GetProjectAbsolutePath([FromQuery] string projectId,
         IProjectFileStructureManager projectFileStructureManager)
     {
         if (string.IsNullOrEmpty(projectId))
@@ -109,7 +47,7 @@ public class ProjectStructureEndpointDefinition : IEndpointDefinition
         return Results.Ok(projectFileStructureManager.GetProjectAbsolutePath(projectId));
     }
 
-    public static async Task<IResult> SetupFileWatcher([FromBody] SetupFileWatcher setupFileWatcher,
+    private static async Task<IResult> SetupFileWatcher([FromBody] SetupFileWatcher setupFileWatcher,
         IValidator<SetupFileWatcher> validator, IProjectFileStructureManager projectFileStructureManager,
         CancellationToken cancellationToken)
     {
@@ -121,15 +59,13 @@ public class ProjectStructureEndpointDefinition : IEndpointDefinition
 
         projectFileStructureManager.SetupFileWatcher(setupFileWatcher.ProjectId);
 
-        // todo return something
-        return Results.Ok("");
+        return Results.Ok();
     }
 
-    public static IResult RemoveFileWatcher(string projectId, IProjectFileStructureManager projectFileStructureManager)
+    private static IResult RemoveFileWatcher(string projectId, IProjectFileStructureManager projectFileStructureManager)
     {
         projectFileStructureManager.RemoveFileWatcher(projectId);
 
-        // todo return something
-        return Results.Ok("");
+        return Results.Ok();
     }
 }
