@@ -1,7 +1,4 @@
-﻿using Microsoft.AspNetCore.Builder;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace ScriptBeeWebApp;
+﻿namespace ScriptBeeWebApp;
 
 public interface IEndpointDefinition
 {
