@@ -1,5 +1,4 @@
-﻿using System.Threading.Tasks;
-using Microsoft.AspNetCore.SignalR;
+﻿using Microsoft.AspNetCore.SignalR;
 using ScriptBeeWebApp.Data;
 using ScriptBeeWebApp.Hubs;
 
