@@ -1,15 +1,10 @@
-﻿using System.Collections.Generic;
-using System.Threading;
-using System.Threading.Tasks;
-using DxWorks.ScriptBee.Plugin.Api.Model;
+﻿using DxWorks.ScriptBee.Plugin.Api.Model;
 using ScriptBee.Models;
 
 namespace ScriptBeeWebApp.Services;
 
 public interface IRunScriptService
 {
-    IEnumerable<string> GetSupportedLanguages();
-
     Task<Run> RunAsync(IProject project, ProjectModel projectModel, string language, string scriptFilePath,
         CancellationToken cancellationToken = default);
 }
