@@ -1,15 +1,17 @@
 ﻿using DxWorks.ScriptBee.Plugin.Api;
+using DxWorks.ScriptBee.Plugin.Api.Model;
 using ScriptBee.Models;
 using ScriptBee.Plugin;
 using ScriptBee.Plugin.Manifest;
 using ScriptBee.ProjectContext;
 using ScriptBee.Services;
 using ScriptBeeWebApp.EndpointDefinitions.Arguments;
+using ScriptBeeWebApp.Repository;
 
 namespace ScriptBeeWebApp.Services;
 
 // todo add tests
-public class LoadersService : ILoadersService
+public sealed class LoadersService : ILoadersService
 {
     private readonly IFileModelService _fileModelService;
     private readonly IPluginRepository _pluginRepository;
@@ -80,40 +82,41 @@ public class LoadersService : ILoadersService
                 continue;
             }
 
-            List<Stream> loadedFileStreams = new();
+            List<NamedFileStream> loadedFileStreams = new();
 
-            foreach (var fileData in loadedFileData)
+            foreach (var fileName in loadedFileData.Select(fileData => fileData.Id.ToString()))
             {
-                var fileStream = await _fileModelService.GetFileAsync(fileData.Id.ToString());
-                loadedFileStreams.Add(fileStream);
+                var fileStream = await _fileModelService.GetFileAsync(fileName);
+                loadedFileStreams.Add(new NamedFileStream(fileName, fileStream));
             }
 
             var dictionary = await modelLoader.LoadModel(loadedFileStreams, cancellationToken: cancellationToken);
 
             _projectManager.AddToGivenProject(projectId, dictionary, modelLoader.GetName());
 
-            AddToLoadedModels(dictionary, loader);
+            AddToLoadedModels(dictionary, loader, loadModels);
 
             foreach (var fileStream in loadedFileStreams)
             {
-                await fileStream.DisposeAsync();
+                await fileStream.Stream.DisposeAsync();
             }
         }
 
         return loadModels;
+    }
 
-        void AddToLoadedModels(Dictionary<string, Dictionary<string, ScriptBeeModel>> dictionary, string loader)
+    private static void AddToLoadedModels(Dictionary<string, Dictionary<string, ScriptBeeModel>> dictionary,
+        string loader, IDictionary<string, List<string>> loadModels)
+    {
+        foreach (var model in dictionary.Keys)
         {
-            foreach (var model in dictionary.Keys)
+            if (loadModels.TryGetValue(model, out var loaders))
             {
-                if (loadModels.TryGetValue(model, out var loaders))
-                {
-                    loaders.Add(loader);
-                }
-                else
-                {
-                    loadModels.Add(model, new List<string> { loader });
-                }
+                loaders.Add(loader);
+            }
+            else
+            {
+                loadModels.Add(model, new List<string> { loader });
             }
         }
     }
